Fix reloading if files are replaced
On linux inotify is tied to the inode, so if the files are
replaced (rather than re-written in place) the previous code
wouldn't be watching the new files.

The fix instead watches the diretory and then reloads when
filenames matching the original names are modified.

Another issue I ran into was the vault agent having a small
window between updating the certificate and key, so the
reloading would always have a noisy "key and cert don't match"
error. Fix this by always waiting 2 seconds after a change
to do the reload.